refactor: Merge types/docs from crud/endpoint.pyi into crud/endpoint.py
- Merged type hints and docstrings from crudclient/crud/endpoint.pyi into crudclient/crud/endpoint.py.

- Deleted crudclient/crud/endpoint.pyi.
